+++ src/commands/should-be-visible.js
@@ -1,6 +1,6 @@
 /**
  * Ensure that the subject is visible.
  */
-Cypress.Commands.add("shouldBeVisible", { prevSubject: "element" }, (subject) => {
+Cypress.Commands.add("shouldBeVisible", { prevSubject: "element" }, subject => {
 	cy.wrap(subject).should("be.visible");
 });