+++ src/commands/should-have-focus.js
@@ -1,6 +1,6 @@
 /**
  * Ensure that the subject has focus.
  */
-Cypress.Commands.add("shouldHaveFocus", { prevSubject: "element" }, (subject) => {
+Cypress.Commands.add("shouldHaveFocus", { prevSubject: "element" }, subject => {
 	cy.wrap(subject).should("have.focus");
 });