+++ src/commands/should-not-be-visible.js
@@ -1,6 +1,6 @@
 /**
  * Ensure that the subject is NOT visible.
  */
-Cypress.Commands.add("shouldNotBeVisible", { prevSubject: "element" }, (subject) => {
+Cypress.Commands.add("shouldNotBeVisible", { prevSubject: "element" }, subject => {
 	cy.wrap(subject).should("not.be.visible");
 });