+++ src/commands/should-not-have-focus.js
@@ -1,6 +1,6 @@
 /**
  * Ensure that the subject does not have focus.
  */
-Cypress.Commands.add("shouldNotHaveFocus", { prevSubject: "element" }, (subject) => {
+Cypress.Commands.add("shouldNotHaveFocus", { prevSubject: "element" }, subject => {
 	cy.wrap(subject).should("not.have.focus");
 });